chore(studio): move functions into utils, move hook
